listening-comp/requirements.txt
youtube-transcript-api>=0.6.2
chromadb>=0.4.22
streamlit>=1.32.0
google-cloud-texttospeech>=2.14.0
groq>=0.4.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
sentence-transformers>=4.0.0
